Added `TERRA_THEME_CONFIG` global to the webpack bundle.
